Side Navigation Bar part 2
Continued working on the side bar. It still needs functionality to
collapse like out top nav bar. Maybe we can make the two bars combine
into on list when on a mobile device. For now its a more visually
appealing solution to a nav bar. I used glyphicons that come with
bootstrap. I also applied it to all of out pages and included the links
to those pages.

I did not delete the code for that other green nav bar. I simply
commented it out.